plugin-samples/abra/gateway/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
pyyaml>=6.0
